CSVEditor: Fix delimiter bug (closes #57)
This bug had a really simple cause: the code to split host and container
assumed that there would be only one delimiter. Now that regex is in
play, it's likely that whatever the delimiter is decided upon, it will
be found in some host string.

The fix is to assume that the container name _doesn't_ have
the delimiter in it, and therefore the last element in the
delimiter-split array is the container and everything before is the
host.