First visible version of PreferenceGroup
Saving hasn't been tested yet!

Additionally the ChoicePreference and StringPreference have been added.

A major change is the introduction of the preferences.json and building
 the preferences UI from a JSON. This allows us to separate the definition
 of preferences from the code as much as possible.

A greater separation of concerns was further introduced using the raw-loader.
It will mark a move away from using the <template> node in index.html
 to a direct import thanks to webpack. This will also help with finding
 the corresponding files for the preference UI more quickly.

A somewhat minor change is the renaming of uiName, ui_title and other
 params to the Preference constructor to simpler and more common names.
Maybe the "ui" prefix will be introduced later on again if it isn't clear
 what the parameter is used for.