Unify container switching conditions
The conditions when opening in the default firefox container (NO_CONTAINER in our code)
 were different to when in any other container. We didn't check if the container was changing,
 so NO_CONTAINER targets would keep closing the old tab and creating a new one despite no
 actual container change happening.
That was because it was assumed passing a container ID when creating a tab would always open the tab a container. In fact the default container does have an ID, thus we can always pass a container ID.

We now do all the same checks before opening a tab in a new container.

Related to #102 - Browsing sites bound to No Container open up in new tabs